requirements.txt
fastapi==0.123.0
pydantic==2.12.5
python-dotenv==1.2.1
uvicorn==0.38.0
requests==2.32.5
mongoengine==0.29.1
slack-sdk==3.39.0
slack-bolt==1.27.0
cryptography==46.0.3